pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-agile-authoring"
version = "0.1.0"
description = "Agile content authoring workflows powered by AI agent skills and modes"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [
    {name = "OSCAL Compass", email = "oscal-compass-oversight@googlegroups.com"}
]
dependencies = ["ruamel.yaml"]

[project.scripts]
agentic-agile-authoring = "agentic_agile_authoring.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_agile_authoring"]

[tool.hatch.build.targets.wheel.force-include]
"skills"                              = "agentic_agile_authoring/data/skills"
"agents-roo/agentic-agile-authoring" = "agentic_agile_authoring/data/roo/agentic-agile-authoring"
".mcp.json"                           = "agentic_agile_authoring/data/mcp.json"
